MUCH MORE FUNCTIONAL! This is the first commit with image and device code, as well as a workspace for multiple ongoing projects that utilize the libraries I'm building.
Serializer code is still a work in progress,
but I've figured out how pySerial did non-blocking IO. There are still
bugs to work on.

There is image processing code to segment a green laser pointer, also
still a work in progress, that will be able to drive the robot. Future
updates will include code to handle control over the network or via an
Xbox 360 wireless controller or Wiimote.