Xbox controller controlling the robot works!
There are still many issues regarding how to handle data read from the
line, but hopefully they will be solved soon. I would like to add
ACK/NACK handling before moving on. (Although it appears that won't be
absolutely necessary since select() is making sure most commands are
written to the line.)

The turning needs to be less sharp (with the Xbox controller), and the
speed needs to now be calculated for 1/2, 1/3, 1/4, etc. speeds. Try to
get it driving accurately. Maybe even add special maneuvers to the
buttons, such as 360 spin, etc.